pyproject.toml
[project]
name = "agent-wallet"
version = "0.1.0"
description = "Model/agent-agnostic shared wallet with per-agent permissions, powered by Coinbase AgentKit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coinbase-agentkit>=0.7.4",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]
